Change signature of Server.getEvents: add pagination.
This also, unfortunately, drops the support for `...filters`,
which means callers need to wrap their filters in a list, e.g.

    getEvents(1, 2, [{ /* EventFilter object here */ }])

This is necessary since we can't have `...filters` and then a
`pagination` object (or parameters), yet putting pagination first
is awkward, both ergonomically and because they're optional.

So the only choice is to lose the convenient `...` expander.