Adds a `ContractSpec` class to parse `ScVal` alongside type info (#115)
This adds a new ContractSpec class which is comprised of the spec entries 
of a contract. This allows the internal user defined types to be accessible 
making the parsing to and from XDR more accurate. This may eventually 
be moved to the js-stellar-base, but for now it is best served here for the 
type support.